dom/webauthn/tests/test_webauthn_make_credential.html

<!DOCTYPE html>
<meta charset=utf-8>
<head>
  <title>Test for MakeCredential for W3C Web Authentication</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="u2futil.js"></script>
  <script type="text/javascript" src="pkijs/common.js"></script>
  <script type="text/javascript" src="pkijs/asn1.js"></script>
  <script type="text/javascript" src="pkijs/x509_schema.js"></script>
  <script type="text/javascript" src="pkijs/x509_simpl.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>

  <h1>Test for MakeCredential for W3C Web Authentication</h1>
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1309284">Mozilla Bug 1309284</a>

  <script class="testbody" type="text/javascript">
   "use strict";

   // Execute the full-scope test
   SimpleTest.waitForExplicitFinish();

   function arrivingHereIsGood(aResult) {
     ok(true, "Good result! Received a: " + aResult);
     return Promise.resolve();
   }

   function arrivingHereIsBad(aResult) {
     ok(false, "Bad result! Received a: " + aResult);
     return Promise.resolve();
   }

   function expectNotAllowedError(aResult) {
     ok(aResult.toString().startsWith("NotAllowedError"), "Expecting a NotAllowedError");
     return Promise.resolve();
   }

   function expectTypeError(aResult) {
     ok(aResult.toString().startsWith("TypeError"), "Expecting a TypeError");
     return Promise.resolve();
   }

   function expectNotSupportedError(aResult) {
     ok(aResult.toString().startsWith("NotSupportedError"), "Expecting a NotSupportedError");
     return Promise.resolve();
   }

   SpecialPowers.pushPrefEnv({"set": [["security.webauth.webauthn", true],
                                      ["security.webauth.webauthn_enable_softtoken", true],
                                      ["security.webauth.webauthn_enable_usbtoken", false]]}, runTests);
   function runTests() {
     isnot(navigator.authentication, undefined, "WebAuthn API endpoint must exist");
     isnot(navigator.authentication.makeCredential, undefined, "WebAuthn makeCredential API endpoint must exist");
     isnot(navigator.authentication.getAssertion, undefined, "WebAuthn getAssertion API endpoint must exist");

     let authn = navigator.authentication;

     let gCredentialChallenge = new Uint8Array(16);
     window.crypto.getRandomValues(gCredentialChallenge);

     let acct = {rpDisplayName: "none", displayName: "none", id: "none"};
     let param = {type: "ScopedCred", algorithm: "p-256"};
     let unsupportedParam = {type: "ScopedCred", algorithm: "3DES"};
     let badParam = {type: "SimplePassword", algorithm: "MaxLength=2"};

     var testFuncs = [
       // Test basic good call
       function() {
         return authn.makeCredential(acct, [param], gCredentialChallenge)
              .then(arrivingHereIsGood)
              .catch(arrivingHereIsBad);
       },

       // Test empty account
       function() {
         return authn.makeCredential({}, [param], gCredentialChallenge)
                     .then(arrivingHereIsBad)
                     .catch(expectTypeError);
       },

       // Test without a parameter
       function() {
         return authn.makeCredential(acct, [], gCredentialChallenge)
                     .then(arrivingHereIsBad)
                     .catch(expectNotSupportedError);
       },

       // Test without a parameter array at all
       function() {
         return authn.makeCredential(acct, null, gCredentialChallenge)
              .then(arrivingHereIsBad)
              .catch(expectTypeError);
       },

       // Test with an unsupported parameter
       function() {
         return authn.makeCredential(acct, [unsupportedParam], gCredentialChallenge)
              .then(arrivingHereIsBad)
              .catch(expectNotSupportedError);
       },

       // Test with an unsupported parameter and a good one
       function() {
         return authn.makeCredential(acct, [unsupportedParam, param], gCredentialChallenge)
              .then(arrivingHereIsGood)
              .catch(arrivingHereIsBad);
       },

       // Test with a bad parameter
       function() {
         return authn.makeCredential(acct, [badParam], gCredentialChallenge)
              .then(arrivingHereIsBad)
              .catch(expectTypeError);
       },

       // Test with an unsupported parameter, and a bad one
       function() {
         return authn.makeCredential(acct, [unsupportedParam, badParam],
                              gCredentialChallenge)
              .then(arrivingHereIsBad)
              .catch(expectTypeError);
       },

       // Test with an unsupported parameter, a bad one, and a good one. This
       // should still fail, as anything with a badParam should fail.
       function() {
         return authn.makeCredential(acct, [unsupportedParam, badParam, param],
                              gCredentialChallenge)
              .then(arrivingHereIsBad)
              .catch(expectTypeError);
       },

       // Test without a challenge
       function() {
         return authn.makeCredential(acct, [param], null)
              .then(arrivingHereIsBad)
              .catch(expectTypeError);
       },

       // Test with an invalid challenge
       function() {
         return authn.makeCredential(acct, [param], "begone, thou ill-fitting moist glove!")
              .then(arrivingHereIsBad)
              .catch(expectTypeError);
       },

       // Test with duplicate parameters
       function() {
         return authn.makeCredential(acct, [param, param, param], gCredentialChallenge)
              .then(arrivingHereIsGood)
              .catch(arrivingHereIsBad);
       },

       // Test an incomplete account
       function() {
         return authn.makeCredential({id: "none"
         }, [param], gCredentialChallenge)
              .then(arrivingHereIsBad)
              .catch(expectTypeError);
       },

       function() {
         return authn.makeCredential({name: "none", imageURL: "http://example.com/404"},
                              [param], gCredentialChallenge)
              .then(arrivingHereIsBad)
              .catch(expectTypeError);
       },

       // Test a complete account
       function() {
         return authn.makeCredential({rpDisplayName: "Foxxy", displayName: "Foxxy V",
                                      id: "foxes_are_the_best@example.com",
                                      name: "Fox F. Foxington",
                                      imageURL: "https://example.com/fox.svg"},
                                     [param], gCredentialChallenge)
                     .then(arrivingHereIsGood)
                     .catch(arrivingHereIsBad);
       }];

     var i = 0;
     var runNextTest = () => {
       if (i == testFuncs.length) {
         SimpleTest.finish();
         return;
       }
       testFuncs[i]().then(() => { runNextTest(); });
       i = i + 1;
     };
     runNextTest();
   };

  </script>

</body>
</html>
